Address linter warnings

+++ atomic.cc
@@ -105,7 +105,7 @@ auto photoionization_crosssection_fromtable(const float *const photoion_xs, cons
     if (nu == nu_edge) {
       sigma_bf = photoion_xs[0];
     } else if (nu <= nu_edge * (1 + globals::NPHIXSNUINCREMENT * globals::NPHIXSPOINTS)) {
-      int const i = floor(nu / (globals::NPHIXSNUINCREMENT * nu_edge)) - 10;
+      int const i = static_cast<int>(floor(nu / (globals::NPHIXSNUINCREMENT * nu_edge)) - 10);
       sigma_bf = photoion_xs[i];
     } else {
       /// use a parameterization of sigma_bf by the Kramers formula
